requirements.txt
hyperliquid-python-sdk>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
pydantic>=2.0.0
loguru>=0.7.0
matplotlib>=3.7.0
